requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0

# openai==1.3.5
# pydantic==1.10.13
# python-multipart==0.0.6
# pydantic-settings==2.1.0
# jinja2==3.1.2
# faker==20.1.0

# langchain-core==0.1.27
# langchain-community==0.0.24
# langchain==0.1.9

# fastapi==0.104.1
# uvicorn==0.24.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0
# openai==1.12.0
# pydantic==2.6.4
# pydantic-core==2.16.2
# pydantic-settings==2.2.1
# python-multipart==0.0.6
# jinja2==3.1.2
# faker==20.1.0

# google-generativeai>=0.3.2


fastapi>=0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
faker==20.1.0
pydantic>=2.3.0
langchain-core>=0.3.52,<0.4.0
langchain-community>=0.0.24
google-generativeai>=0.3.2
google-ai-generativelanguage>=0.6.18,<0.7.0
langchain-google-genai>=2.1.4
pydantic-settings>=2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0